refactor: Updated internal imports of crop split/merge
